changed variable to track number of recursive calls in the max clique method to be a long. It is now clear that for instances of this problem, the number of recursive calls will exceed MAXINT.

+++ src/main/java/com/aaronpmaus/jMath/graph/UndirectedGraph.java
@@ -14,7 +14,7 @@ import java.util.Date;
  * @since 0.1.0
 */
 public class UndirectedGraph<T> extends Graph<T>{
-    public static int numRecursiveCalls = 0;
+    public static long numRecursiveCalls = 0;
 
     /**
      * A default constructor for the UndirectedGraph
